+++ lyscripts/__init__.py
@@ -9,19 +9,7 @@ from rich_argparse import RichHelpFormatter
 
 from lyscripts.utils import report
 
-from . import (
-    clean,
-    enhance,
-    evaluate,
-    generate,
-    join,
-    plot,
-    predict,
-    sample,
-    split,
-    temp_schedule,
-    utils,
-)
+from . import app, data, evaluate, plot, predict, sample, temp_schedule, utils
 from ._version import version
 
 __version__ = version
@@ -97,15 +85,12 @@ def main():
 
     # the individual scripts add `ArgumentParser` instances and their arguments to
     # this `subparsers` object
-    generate._add_parser(subparsers, help_formatter=parser.formatter_class)
-    join._add_parser(subparsers, help_formatter=parser.formatter_class)
-    enhance._add_parser(subparsers, help_formatter=parser.formatter_class)
-    clean._add_parser(subparsers, help_formatter=parser.formatter_class)
-    split._add_parser(subparsers, help_formatter=parser.formatter_class)
-    sample._add_parser(subparsers, help_formatter=parser.formatter_class)
+    app._add_parser(subparsers, help_formatter=parser.formatter_class)
+    data._add_parser(subparsers, help_formatter=parser.formatter_class)
     evaluate._add_parser(subparsers, help_formatter=parser.formatter_class)
-    predict._add_parser(subparsers, help_formatter=parser.formatter_class)
     plot._add_parser(subparsers, help_formatter=parser.formatter_class)
+    predict._add_parser(subparsers, help_formatter=parser.formatter_class)
+    sample._add_parser(subparsers, help_formatter=parser.formatter_class)
     temp_schedule._add_parser(subparsers, help_formatter=parser.formatter_class)
 
     args = parser.parse_args()