+++ lyscripts/data/__main__.py
@@ -1,12 +1,12 @@
 import argparse
 
 from lyscripts import RichDefaultHelpFormatter, exit_cli
-from lyscripts.predict import prevalences, risks
+from lyscripts.data import clean, enhance, generate, join, split
 
 # I need another __main__ guard here, because otherwise pdoc tries to run this
 if __name__ == "__main__":
     parser = argparse.ArgumentParser(
-        prog="lyscripts predict",
+        prog="lyscripts data",
         description=__doc__,
         formatter_class=RichDefaultHelpFormatter,
     )
@@ -15,8 +15,11 @@ if __name__ == "__main__":
 
     # the individual scripts add `ArgumentParser` instances and their arguments to
     # this `subparsers` object
-    risks._add_parser(subparsers, help_formatter=parser.formatter_class)
-    prevalences._add_parser(subparsers, help_formatter=parser.formatter_class)
+    clean._add_parser(subparsers, help_formatter=parser.formatter_class)
+    enhance._add_parser(subparsers, help_formatter=parser.formatter_class)
+    generate._add_parser(subparsers, help_formatter=parser.formatter_class)
+    join._add_parser(subparsers, help_formatter=parser.formatter_class)
+    split._add_parser(subparsers, help_formatter=parser.formatter_class)
 
     args = parser.parse_args()
     args.run_main(args)