+++ lyscripts/data/generate.py
@@ -7,7 +7,8 @@ from pathlib import Path
 import emcee
 import numpy as np
 
-from lyscripts.utils import cli_load_yaml_params, model_from_config, report
+from lyscripts.data.utils import save_table_to_csv
+from lyscripts.utils import create_model_from_config, load_yaml_params, report
 
 
 def _add_parser(
@@ -68,42 +69,34 @@ def main(args: argparse.Namespace):
     The CLI's help for this subcommand (`lyscripts generate --help`) shows:
 
     ```
-    usage: lyscripts generate [-h] [--params PARAMS]
-                            [--set-theta SET_THETA [SET_THETA ...] | --load-theta
-                            {mean,max_llh}] [--samples SAMPLES]
-                            num output
+    USAGE: lyscripts data generate [-h] [--params PARAMS]
+                                   [--set-theta SET_THETA [SET_THETA ...] |
+                                   --load-theta {mean,max_llh}] [--samples SAMPLES]
+                                   num output
 
-    Generate synthetic patient data for testing purposes.
+    Generate synthetic patient data for testing and validation purposes.
 
+    POSITIONAL ARGUMENTS:
+      num                   Number of synthetic patient records to generate
+      output                Path where to store the generated synthetic data
 
-    POSITIONAL ARGUMENTS
-    num                                   Number of synthetic patient records to
-                                            generate
-    output                                Path where to store the generated synthetic
-                                            data
-
-    OPTIONAL ARGUMENTS
-    -h, --help                            show this help message and exit
-    --params PARAMS                       Parameter file containing model specifications
-                                            (default: ./params.yaml)
-    --set-theta SET_THETA [SET_THETA      Set the spread probs and parameters for time
-    ...]                                  marginalization by hand (default: None)
-    --load-theta {mean,max_llh}           Use either the mean or the maximum likelihood
-                                            estimate from drawn samples (default: mean)
-    --samples SAMPLES                     Path to the samples if a method to load them
-                                            was chosen (default: ./models/samples.hdf5)
+    OPTIONAL ARGUMENTS:
+      -h, --help            show this help message and exit
+      --params PARAMS       Parameter file containing model specifications (default:
+                            ./params.yaml)
+      --set-theta SET_THETA [SET_THETA ...]
+                            Set the spread probs and parameters for time
+                            marginalization by hand (default: None)
+      --load-theta {mean,max_llh}
+                            Use either the mean or the maximum likelihood estimate
+                            from drawn samples (default: mean)
+      --samples SAMPLES     Path to the samples if a method to load them was chosen
+                            (default: ./models/samples.hdf5)
     ```
     """
-    params = cli_load_yaml_params(args.params)
-
-    with report.status("Create model..."):
-        model = model_from_config(
-            graph_params=params["graph"],
-            model_params=params["model"],
-            modalities_params=params["modalities"],
-        )
-        ndim = len(model.spread_probs) + model.diag_time_dists.num_parametric
-        report.success(f"Created {type(model)} model")
+    params = load_yaml_params(args.params)
+    model = create_model_from_config(params)
+    ndim = len(model.spread_probs) + model.diag_time_dists.num_parametric
 
     if args.set_theta is not None:
         with report.status("Assign given parameters to model..."):
@@ -111,9 +104,9 @@ def main(args: argparse.Namespace):
                 raise ValueError(
                     f"Model takes {ndim} parameters, but{len(args.set_theta)} were provided"
                 )
-            THETA = np.array(args.set_theta)
-            model.check_and_assign(THETA)
-            report.print(THETA)
+            theta = np.array(args.set_theta)
+            model.check_and_assign(theta)
+            report.print(theta)
             report.success("Assigned given parameters to model")
     else:
         with report.status(f"Load samples and choose their {args.load_theta} value..."):
@@ -126,15 +119,15 @@ def main(args: argparse.Namespace):
             log_probs = backend.get_blobs(flat=True)
 
             if args.load_theta == "mean":
-                THETA = np.mean(chain, axis=0)
+                theta = np.mean(chain, axis=0)
             elif args.load_theta == "max_llh":
                 max_llh_idx = np.argmax(log_probs)
-                THETA = chain[max_llh_idx]
+                theta = chain[max_llh_idx]
             else:
                 raise ValueError("Only 'mean' and 'max_llh' are supported")
 
-            model.check_and_assign(THETA)
-            report.print(THETA)
+            model.check_and_assign(theta)
+            report.print(theta)
             report.success(f"Loaded samples and assigned their {args.load_theta} value")
 
     with report.status(f"Generate synthetic data of {args.num} patients..."):
@@ -150,10 +143,7 @@ def main(args: argparse.Namespace):
             )
         report.success(f"Created synthetic data of {args.num} patients.")
 
-    with report.status("Save generated dataset..."):
-        args.output.parent.mkdir(exist_ok=True)
-        synthetic_data.to_csv(args.output, index=None)
-        report.success(f"Saved generated dataset to {args.output}")
+    save_table_to_csv(args.output, synthetic_data)
 
 
 if __name__ == "__main__":