+++ lyscripts/data/split.py
@@ -9,7 +9,7 @@ from pathlib import Path
 import numpy as np
 import pandas as pd
 
-from lyscripts.utils import cli_load_yaml_params, report
+from lyscripts.utils import load_yaml_params, report
 
 warnings.simplefilter(action="ignore", category=FutureWarning)
 
@@ -57,22 +57,21 @@ def main(args: argparse.Namespace):
     The output from `lyscripts split --help` looks like this:
 
     ```
-    usage: lyscripts split [-h] [-p PARAMS] input output
+    USAGE: lyscripts data split [-h] [-p PARAMS] input output
 
     Split the full dataset into cross-validation folds according to the content of the
     params.yaml file.
 
+    POSITIONAL ARGUMENTS:
+      input                 The path to the full dataset to split.
+      output                Folder to store the split CSV files in.
 
-    POSITIONAL ARGUMENTS
-    input                The path to the full dataset to split.
-    output               Folder to store the split CSV files in.
-
-    OPTIONAL ARGUMENTS
-    -h, --help           show this help message and exit
-    -p, --params PARAMS  Path to parameter YAML file. (default: ./params.yaml)
+    OPTIONAL ARGUMENTS:
+      -h, --help            show this help message and exit
+      -p, --params PARAMS   Path to parameter YAML file. (default: ./params.yaml)
     ```
     """
-    params = cli_load_yaml_params(args.params)
+    params = load_yaml_params(args.params)
 
     with report.status("Reading in concatenated CSV file..."):
         header = [0,1] if params["model"]["class"] == "Unilateral" else [0,1,2]