+++ lyscripts/plot/utils.py
@@ -13,7 +13,11 @@ import scipy as sp
 from matplotlib.axes._axes import Axes as MPLAxes
 from matplotlib.figure import Figure
 
-from lyscripts.utils import report
+from lyscripts.utils import (
+    check_input_file_exists,
+    check_output_dir_exists,
+    report_state,
+)
 
 # define USZ colors
 COLORS = {
@@ -71,18 +75,22 @@ def _clean_and_check(filename: Union[str, Path]) -> Path:
 @dataclass
 class Histogram:
     """Class containing data for plotting a histogram."""
-    filename: Union[str, Path]
-    dataname: str
+    values: np.ndarray
     scale: float = field(default=100.)
     kwargs: Dict[str, Any] = field(default_factory=lambda: {})
 
     def __post_init__(self) -> None:
-        self.filename = _clean_and_check(self.filename)
-        with h5py.File(self.filename, mode="r") as h5file:
-            dataset = h5file[self.dataname]
-            self.values = self.scale * dataset[:]
-            if "label" not in self.kwargs:
-                self.kwargs["label"] = get_label(dataset.attrs)
+        self.values = self.scale * self.values
+
+    @classmethod
+    def from_hdf5(cls, filename, dataname, scale: float = 100., **kwargs):
+        """Create a histogram from an HDF5 file."""
+        filename = _clean_and_check(filename)
+        with h5py.File(filename, mode="r") as h5file:
+            dataset = h5file[dataname]
+            if "label" not in kwargs:
+                kwargs["label"] = get_label(dataset.attrs)
+            return cls(values=dataset[:], scale=scale, kwargs=kwargs)
 
     def left_percentile(self, percent: float) -> float:
         """Compute the point where `percent` of the values are to the left."""
@@ -95,23 +103,30 @@ class Histogram:
 @dataclass
 class Posterior:
     """Class for storing plot configs for a Beta posterior."""
-    filename: Union[str, Path]
-    dataname: str
+    num_success: int
+    num_total: int
     scale: float = field(default=100.)
     kwargs: Dict[str, Any] = field(default_factory=lambda: {})
 
-    def __post_init__(self) -> None:
-        self.filename = _clean_and_check(self.filename)
-        with h5py.File(self.filename, mode="r") as h5file:
-            dataset = h5file[self.dataname]
+    @classmethod
+    def from_hdf5(cls, filename, dataname, scale: float = 100., **kwargs) -> None:
+        """Initialize data container for Beta posteriors from HDF5 file."""
+        filename = _clean_and_check(filename)
+        with h5py.File(filename, mode="r") as h5file:
+            dataset = h5file[dataname]
             try:
-                self.num_success = int(dataset.attrs["num_match"])
-                self.num_total = int(dataset.attrs["num_total"])
+                num_success = int(dataset.attrs["num_match"])
+                num_total = int(dataset.attrs["num_total"])
             except KeyError as key_err:
                 raise KeyError(
                     "Dataset does not contain observed prevalence data"
                 ) from key_err
-            self.num_fail = self.num_total - self.num_success
+
+        return cls(num_success, num_total, scale=scale, kwargs=kwargs)
+
+    @property
+    def num_fail(self):
+        return self.num_total - self.num_success
 
     def pdf(self, x: np.ndarray) -> np.ndarray:
         """Compute the probability density function."""
@@ -260,33 +275,26 @@ def draw(
     return axes
 
 
+@report_state(
+    status_msg="Load MPL stylesheet...",
+    success_msg="Loaded MPL stylesheet.",
+)
+@check_input_file_exists
 def use_mpl_stylesheet(file_path: Union[str, Path]):
-    """
-    Attempt to load a `.mplstyle` stylesheet from `file_path` and fail gracefully if it
-    is not found.
-    """
-    file_path = Path(file_path)
-    with report.status("Apply MPL stylesheet..."):
-        try:
-            plt.style.use(file_path)
-        except OSError:
-            report.failure("Could not find MPL stylesheet.")
-        else:
-            report.success(f"Applied MPL stylesheet from {file_path}")
+    """Load a `.mplstyle` stylesheet from `file_path`."""
+    plt.style.use(file_path)
 
 
+@report_state(
+    status_msg="Save matplotlib figure...",
+    success_msg="Saved matplotlib figure.",
+)
+@check_output_dir_exists
 def save_figure(
-    figure: Figure,
     output_path: Union[str, Path],
+    figure: Figure,
     formats: Optional[List[str]],
 ):
-    """
-    Save a `figure` at `output_path` in every one of the provided `formats`. If not
-    `formats` are provided, the default is `PNG` and `SVG`.
-    """
-    output_path = Path(output_path)
-    output_path.parent.mkdir(exist_ok=True)
-    with report.status("Save figure..."):
-        for frmt in formats:
-            figure.savefig(output_path.with_suffix(f".{frmt}"))
-        report.success(f"Saved figure to {output_path} in the formats {formats}.")
+    """Save a `figure` to `output_path` in every one of the provided `formats`."""
+    for frmt in formats:
+        figure.savefig(output_path.with_suffix(f".{frmt}"))